fix: reconsolidated IPv6 util functions into NetAddr file
