+++ packages/core/src/utils/const/index.ts
@@ -1,3 +1,2 @@
 export * from './abi';
 export * from './data';
-export * from './ip';
