+++ packages/core/src/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './const';
 export * from './data';
-export * from './ip';
