--- packages/core/src/utils/ip/index.ts
@@ -1,1 +1,0 @@
-export * from './ip';
